backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
pydantic
python-jose
passlib[bcrypt]
redis
celery
psycopg2-binary